Restore identification of HP VDF files as application/x-hp-vdf
Recent re-work of the sql magic rules (probably as a result of merging
upstream changes) caused the VDF files to be reported as the generic
application/x-sqlite3 type. It appears the first MIME type assigned to
a file "wins", so change the sqlite3 rules not to apply the x-sqlite3
type until after checking more specific types.